requirements.txt
requests>=2.31
SQLAlchemy>=2.0
pyodbc>=5.0
python-dotenv>=1.0
